@yosiat 2016-03-19
By the way, I think this pull request and #356 will solve the high cpu usage issue - #267 ,
I don't know how to compile and rpm to check this out.