@nathanielc 2016-03-21
@yosiat This is great work thanks. I am going to play around with this locally to see if I can't reduce some of the overhead introduced for the 10 task case.