@yosiat 2016-03-21
@nathanielc I think that the overhead is one the creation of the tasks and not running the data.