@nathanielc 2016-03-21
@yosiat I thought the benchmark didn't time task creation anymore?