@yosiat 2016-03-22
@nathanielc After doing some research, I ran the profiling for matches and not matches only for 10 points it looks that most of the time (1.7s) is on ioutil.ReadAll on serveWrite.

Here is the result for top50 on not matches profile:
```
      flat  flat%   sum%        cum   cum%
     1.59s 61.39% 61.39%      1.59s 61.39%  syscall.Syscall
     0.31s 11.97% 73.36%      0.33s 12.74%  runtime.scanobject
     0.10s  3.86% 77.22%      0.10s  3.86%  runtime.memmove
     0.10s  3.86% 81.08%      0.10s  3.86%  runtime.usleep
     0.05s  1.93% 83.01%      0.05s  1.93%  runtime.deductSweepCredit
     0.05s  1.93% 84.94%      0.05s  1.93%  runtime.mach_semaphore_signal
     0.05s  1.93% 86.87%      0.15s  5.79%  runtime.mallocgc
     0.04s  1.54% 88.42%      0.04s  1.54%  nanotime
     0.04s  1.54% 89.96%      0.04s  1.54%  runtime.mach_semaphore_timedwait
     0.03s  1.16% 91.12%      0.03s  1.16%  runtime.mach_semaphore_wait
     0.02s  0.77% 91.89%      0.02s  0.77%  runtime.(*mcentral).grow
     0.02s  0.77% 92.66%      0.02s  0.77%  runtime.funcspdelta
     0.02s  0.77% 93.44%      0.45s 17.37%  runtime.gcDrain
     0.02s  0.77% 94.21%      0.05s  1.93%  runtime.mapassign1
     0.01s  0.39% 94.59%      1.88s 72.59%  github.com/influxdata/kapacitor/services/httpd.(*Handler).ServeHTTP
     0.01s  0.39% 94.98%      0.02s  0.77%  runtime.makeslice
     0.01s  0.39% 95.37%      0.02s  0.77%  strconv.formatBits
     0.01s  0.39% 95.75%      0.02s  0.77%  strings.Join
         0     0% 95.75%      1.59s 61.39%  bytes.(*Buffer).ReadFrom
         0     0% 95.75%      1.88s 72.59%  command-line-arguments.BenchWrite
         0     0% 95.75%      1.88s 72.59%  command-line-arguments.Benchmark_Write_MeasurementNameNotMatches_10
         0     0% 95.75%      0.05s  1.93%  github.com/influxdata/influxdb/models.(*point).Fields
         0     0% 95.75%      0.05s  1.93%  github.com/influxdata/influxdb/models.(*point).unmarshalBinary
         0     0% 95.75%      0.05s  1.93%  github.com/influxdata/influxdb/models.newFieldsFromBinary
         0     0% 95.75%      0.03s  1.16%  github.com/influxdata/kapacitor.(*Edge).CollectPoint
         0     0% 95.75%      0.08s  3.09%  github.com/influxdata/kapacitor.(*TaskMaster).WritePoints
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.(*Handler).(github.com/influxdata/kapacitor/services/httpd.serveWrite)-fm
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.(*Handler).serveWrite
         0     0% 95.75%      0.09s  3.47%  github.com/influxdata/kapacitor/services/httpd.(*Handler).serveWriteLine
         0     0% 95.75%      1.87s 72.20%  github.com/influxdata/kapacitor/services/httpd.(*ServeMux).ServeHTTP
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.authenticate.func1
         0     0% 95.75%      0.11s  4.25%  github.com/influxdata/kapacitor/services/httpd.buildLogLine
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.cors.func1
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.gzipFilter.func1
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.jsonContent.func1
         0     0% 95.75%      1.86s 71.81%  github.com/influxdata/kapacitor/services/httpd.logging.func1
         0     0% 95.75%      1.87s 72.20%  github.com/influxdata/kapacitor/services/httpd.recovery.func1
         0     0% 95.75%      1.73s 66.80%  github.com/influxdata/kapacitor/services/httpd.requestID.func1
         0     0% 95.75%      1.70s 65.64%  github.com/influxdata/kapacitor/services/httpd.versionHeader.func1
         0     0% 95.75%      1.60s 61.78%  io/ioutil.ReadAll
         0     0% 95.75%      1.60s 61.78%  io/ioutil.readAll
         0     0% 95.75%      1.87s 72.20%  net/http.HandlerFunc.ServeHTTP
         0     0% 95.75%      0.02s  0.77%  net/http.Header.Set
         0     0% 95.75%      0.02s  0.77%  net/textproto.MIMEHeader.Set
         0     0% 95.75%      1.59s 61.39%  os.(*File).Read
         0     0% 95.75%      1.59s 61.39%  os.(*File).read
         0     0% 95.75%      0.09s  3.47%  runtime.(*gcWork).balance
         0     0% 95.75%      0.07s  2.70%  runtime.(*mcache).refill
         0     0% 95.75%      0.07s  2.70%  runtime.(*mcentral).cacheSpan
         0     0% 95.75%      0.02s  0.77%  runtime.convT2E
```

I find it very strange.. what do you say?