@yosiat 2016-03-22
@nathanielc I am seeing something positive but weird happens here ;)
After rebasing the "fork-by-measurement" branch to the latest master, the results are decreasing from 23,000ns~ to 9,000ns~ (thanks to allocations dropping),

Those are the final results:
```
benchmark                                            old ns/op      new ns/op      delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     8633314476     9939           -100.00%
Benchmark_Write_MeasurementNameMatches_1000-4        7915678886     6755653114     -14.65%
Benchmark_Write_MeasurementNameNotMatches_100-4      37434          8926           -76.16%
Benchmark_Write_MeasurementNameMatches_100-4         38474          20700          -46.20%
Benchmark_Write_MeasurementNameNotMatches_10-4       22950          10169          -55.69%
Benchmark_Write_MeasurementNameMatches_10-4          23109          10856          -53.02%

benchmark                                            old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     57450          21             -99.96%
Benchmark_Write_MeasurementNameMatches_1000-4        57426          57407          -0.03%
Benchmark_Write_MeasurementNameNotMatches_100-4      49             21             -57.14%
Benchmark_Write_MeasurementNameMatches_100-4         49             22             -55.10%
Benchmark_Write_MeasurementNameNotMatches_10-4       49             21             -57.14%
Benchmark_Write_MeasurementNameMatches_10-4          49             21             -57.14%

benchmark                                            old bytes     new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     4264608       1978          -99.95%
Benchmark_Write_MeasurementNameMatches_1000-4        4261568       4260176       -0.03%
Benchmark_Write_MeasurementNameNotMatches_100-4      3889          1978          -49.14%
Benchmark_Write_MeasurementNameMatches_100-4         3889          2035          -47.67%
Benchmark_Write_MeasurementNameNotMatches_10-4       3838          1997          -47.97%
Benchmark_Write_MeasurementNameMatches_10-4          3838          1997          -47.97%
```

This is a very dramatic change... do you know what cause it?