@nathanielc 2016-03-22
@yosiat Yes, it turns out that the HTTP logging is very expensive. I found that in smaller tests the cost  to do the HTTP side of things was most expensive so I dug in. A pushed a change to master the turns off HTTP logging entirely if the logging level is such that it wouldn't be logged.