@yosiat 2016-03-22
@nathanielc This is very cool! but it's weird because in my cpu profile I didn't saw that logging takes most of the time, it was complaining about the ioutil.ReadAll in serveWrite..

What do you think about this PR after this improvement? there is something to change?