@yosiat 2016-03-22
@nathanielc Yes, here are the results:
```
benchmark                                            old ns/op      new ns/op      delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     8575928035     13803          -100.00%
Benchmark_Write_MeasurementNameMatches_1000-4        9441685070     6754843321     -28.46%
Benchmark_Write_MeasurementNameNotMatches_100-4      22958          9534           -58.47%
Benchmark_Write_MeasurementNameMatches_100-4         33692          22269          -33.90%
Benchmark_Write_MeasurementNameNotMatches_10-4       11710          12256          +4.66%
Benchmark_Write_MeasurementNameMatches_10-4          10731          12955          +20.73%

benchmark                                            old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     57419          21             -99.96%
Benchmark_Write_MeasurementNameMatches_1000-4        57412          57412          +0.00%
Benchmark_Write_MeasurementNameNotMatches_100-4      22             21             -4.55%
Benchmark_Write_MeasurementNameMatches_100-4         22             22             +0.00%
Benchmark_Write_MeasurementNameNotMatches_10-4       21             21             +0.00%
Benchmark_Write_MeasurementNameMatches_10-4          21             21             +0.00%

benchmark                                            old bytes     new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_1000-4     4262336       1978          -99.95%
Benchmark_Write_MeasurementNameMatches_1000-4        4260480       4260496       +0.00%
Benchmark_Write_MeasurementNameNotMatches_100-4      2035          1978          -2.80%
Benchmark_Write_MeasurementNameMatches_100-4         2086          2035          -2.44%
Benchmark_Write_MeasurementNameNotMatches_10-4       1997          1997          +0.00%
Benchmark_Write_MeasurementNameMatches_10-4          1997          1997          +0.00%
```


The big increase is because I changed another thing to give a big performance boost (and this is the biggest among all) - The old code checked against all forks (in our case 1,000 forks) if they are match which is O(n) complexity.

The new code create a key which is composed by db, rp and measurement and finds all forks by this key - so this is map lookup to find all forks - O(1).
Look at the diff for forkPoint in task_master and you will see the big difference.

Thanks to this change we will be faster on 100,000 tasks with no matches as well.