@yosiat 2016-03-25
@nathanielc I have researched the increase **Benchmark_Write_MeasurementNameMatches_10-4** and tried to run this many times on my computer, because I noticed once the master performance was worse than this pr, and I found times that the pr was slower from the master.

So I tried to run them multiple times on the master:
```
> go test -v -bench="(.*)NameMatches_10$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go 
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameMatches_10-4	  200000	      8407 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	3.063s
> go test -v -bench="(.*)Matches_10$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go 
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameNotMatches_10-4	  200000	      8508 ns/op	    1978 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_10-4   	  200000	      8438 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	4.835s
> go test -v -bench="(.*)Matches_10\d?$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameNotMatches_100-4	  100000	     12878 ns/op	    1997 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_100-4   	  100000	     12641 ns/op	    1997 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameNotMatches_10-4 	  200000	      8563 ns/op	    1978 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_10-4    	  200000	      8436 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	23.940s
```

And on the fork-by-measurement branch:
```
> go test -v -bench="(.*)NameMatches_10$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go 
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameMatches_10-4	  200000	      8772 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	3.164s
> go test -v -bench="(.*)Matches_10$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go 
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameNotMatches_10-4	  200000	      8389 ns/op	    1978 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_10-4   	  200000	      8959 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	4.934s
> go test -v -bench="(.*)Matches_10\d?$" ./integrations/write_benchmark_test.go ./integrations/helpers_test.go
testing: warning: no tests to run
PASS
Benchmark_Write_MeasurementNameNotMatches_100-4	  200000	      8129 ns/op	    1978 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_100-4   	  100000	     11922 ns/op	    1997 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameNotMatches_10-4 	  200000	      8436 ns/op	    1978 B/op	      21 allocs/op
Benchmark_Write_MeasurementNameMatches_10-4    	  200000	      8527 ns/op	    1978 B/op	      21 allocs/op
ok  	command-line-arguments	13.036s
```

And this is final comparison:
```
benchmark                                           old ns/op     new ns/op     delta
Benchmark_Write_MeasurementNameNotMatches_100-4     12878         8129          -36.88%
Benchmark_Write_MeasurementNameMatches_100-4        12641         11922         -5.69%
Benchmark_Write_MeasurementNameNotMatches_10-4      8563          8436          -1.48%
Benchmark_Write_MeasurementNameMatches_10-4         8436          8527          +1.08%

benchmark                                           old allocs     new allocs     delta
Benchmark_Write_MeasurementNameNotMatches_100-4     21             21             +0.00%
Benchmark_Write_MeasurementNameMatches_100-4        21             21             +0.00%
Benchmark_Write_MeasurementNameNotMatches_10-4      21             21             +0.00%
Benchmark_Write_MeasurementNameMatches_10-4         21             21             +0.00%

benchmark                                           old bytes     new bytes     delta
Benchmark_Write_MeasurementNameNotMatches_100-4     1997          1978          -0.95%
Benchmark_Write_MeasurementNameMatches_100-4        1997          1997          +0.00%
Benchmark_Write_MeasurementNameNotMatches_10-4      1978          1978          +0.00%
Benchmark_Write_MeasurementNameMatches_10-4         1978          1978          +0.00%
```

I don't think that 1% increase a 100ns increase worth researching.

About the **Benchmark_Write_MeasurementNameMatches_1000-4** I think this PR shows a big improvement - 28%, and we need to dig this further but I think this for another issue/pr, I think the problem their requires a bigger change.