@yosiat 2016-03-30
@nathanielc I see something really weird after rebasing this branch with the latest benchmarks -
```
benchmark                               old ns/op       new ns/op       delta
Benchmark_T1000_P5000_Matches-4         2573679111      2876176010      +11.75%
Benchmark_T1000_P5000_NoMatches-4       2516031739      2893322930      +15.00%
Benchmark_T100_P5000_Matches-4          30550621        30032541        -1.70%
Benchmark_T100_P5000_NoMatches-4        29795228        31743357        +6.54%
Benchmark_T10_P5000_Matches-4           3524056         3108466         -11.79%
Benchmark_T10_P5000_NoMatches-4         3480425         3398988         -2.34%
Benchmark_T10_P500_CountTask-4          4048433         4051221         +0.07%
Benchmark_T10_P500000_CountTask-4       168250231       166392252       -1.10%
Benchmark_T1000_P500-4                  17615494351     17051748760     -3.20%
Benchmark_T1000_P500000_CountTask-4     19578823592     19163480475     -2.12%

benchmark                               old allocs     new allocs     delta
Benchmark_T1000_P5000_Matches-4         12083          12084          +0.01%
Benchmark_T1000_P5000_NoMatches-4       12084          12080          -0.03%
Benchmark_T100_P5000_Matches-4          1265           1264           -0.08%
Benchmark_T100_P5000_NoMatches-4        1265           1265           +0.00%
Benchmark_T10_P5000_Matches-4           179            178            -0.56%
Benchmark_T10_P5000_NoMatches-4         177            178            +0.56%
Benchmark_T10_P500_CountTask-4          389            389            +0.00%
Benchmark_T10_P500000_CountTask-4       431            428            -0.70%
Benchmark_T1000_P500-4                  32191          32189          -0.01%
Benchmark_T1000_P500000_CountTask-4     32240          32242          +0.01%

benchmark                               old bytes     new bytes     delta
Benchmark_T1000_P5000_Matches-4         1818368       1818368       +0.00%
Benchmark_T1000_P5000_NoMatches-4       1818352       1818048       -0.02%
Benchmark_T100_P5000_Matches-4          1413790       1413721       -0.00%
Benchmark_T100_P5000_NoMatches-4        1413765       1413795       +0.00%
Benchmark_T10_P5000_Matches-4           1373078       1372982       -0.01%
Benchmark_T10_P5000_NoMatches-4         1372951       1372997       +0.00%
Benchmark_T10_P500_CountTask-4          157327        157388        +0.04%
Benchmark_T10_P500000_CountTask-4       149610839     149610832     -0.00%
Benchmark_T1000_P500-4                  1434032       1433936       -0.01%
Benchmark_T1000_P500000_CountTask-4     150886616     150886184     -0.00%
```

It looks likes the fork by measurement didn't improved the non-matches.. this is really weird..