@yosiat 2016-03-30
@nathanielc and those are the results with edge buffer size of 0:
```
benchmark                               old ns/op       new ns/op       delta
Benchmark_T1000_P5000_Matches-4         2505622528      2526439431      +0.83%
Benchmark_T1000_P5000_NoMatches-4       2624906171      2429481570      -7.45%
Benchmark_T100_P5000_Matches-4          41977088        33112785        -21.12%
Benchmark_T100_P5000_NoMatches-4        37774836        33507514        -11.30%
Benchmark_T10_P5000_Matches-4           3819284         3804576         -0.39%
Benchmark_T10_P5000_NoMatches-4         3900131         3887403         -0.33%
Benchmark_T10_P500_CountTask-4          3998147         4683147         +17.13%
Benchmark_T10_P500000_CountTask-4       174010906       163730995       -5.91%
Benchmark_T1000_P500-4                  17339363380     16875495010     -2.68%
Benchmark_T1000_P500000_CountTask-4     17059904179     17326672804     +1.56%

benchmark                               old allocs     new allocs     delta
Benchmark_T1000_P5000_Matches-4         12085          12084          -0.01%
Benchmark_T1000_P5000_NoMatches-4       12088          12084          -0.03%
Benchmark_T100_P5000_Matches-4          1272           1272           +0.00%
Benchmark_T100_P5000_NoMatches-4        1272           1271           -0.08%
Benchmark_T10_P5000_Matches-4           179            179            +0.00%
Benchmark_T10_P5000_NoMatches-4         179            179            +0.00%
Benchmark_T10_P500_CountTask-4          386            387            +0.26%
Benchmark_T10_P500000_CountTask-4       431            432            +0.23%
Benchmark_T1000_P500-4                  32187          32184          -0.01%
Benchmark_T1000_P500000_CountTask-4     32243          32244          +0.00%

benchmark                               old bytes     new bytes     delta
Benchmark_T1000_P5000_Matches-4         1818432       1818368       -0.00%
Benchmark_T1000_P5000_NoMatches-4       1818752       1818384       -0.02%
Benchmark_T100_P5000_Matches-4          1414397       1414272       -0.01%
Benchmark_T100_P5000_NoMatches-4        1414332       1414289       -0.00%
Benchmark_T10_P5000_Matches-4           1373126       1373113       -0.00%
Benchmark_T10_P5000_NoMatches-4         1373108       1373109       +0.00%
Benchmark_T10_P500_CountTask-4          157195        157380        +0.12%
Benchmark_T10_P500000_CountTask-4       149610969     149611025     +0.00%
Benchmark_T1000_P500-4                  1433776       1433664       -0.01%
Benchmark_T1000_P500000_CountTask-4     150887648     150887360     -0.00%
```