@yosiat 2016-03-30
By the way - if we remove the sort function in the updateKeys, we get an high performance boost..
Here is the diff between the fork-by-master and the removal of this:
```
benchmark                               old ns/op       new ns/op      delta
Benchmark_T1000_P5000_Matches-4         2526439431      284995809      -88.72%
Benchmark_T1000_P5000_NoMatches-4       2429481570      298204769      -87.73%
Benchmark_T100_P5000_Matches-4          33112785        18386522       -44.47%
Benchmark_T100_P5000_NoMatches-4        33507514        18454186       -44.93%
Benchmark_T10_P5000_Matches-4           3804576         4395114        +15.52%
Benchmark_T10_P5000_NoMatches-4         3887403         4493374        +15.59%
Benchmark_T10_P500_CountTask-4          4683147         3962128        -15.40%
Benchmark_T10_P500000_CountTask-4       163730995       162851625      -0.54%
Benchmark_T1000_P500-4                  16875495010     1445119778     -91.44%
Benchmark_T1000_P500000_CountTask-4     17326672804     1781148362     -89.72%

benchmark                               old allocs     new allocs     delta
Benchmark_T1000_P5000_Matches-4         12084          10082          -16.57%
Benchmark_T1000_P5000_NoMatches-4       12084          10068          -16.68%
Benchmark_T100_P5000_Matches-4          1272           1064           -16.35%
Benchmark_T100_P5000_NoMatches-4        1271           1064           -16.29%
Benchmark_T10_P5000_Matches-4           179            156            -12.85%
Benchmark_T10_P5000_NoMatches-4         179            156            -12.85%
Benchmark_T10_P500_CountTask-4          387            333            -13.95%
Benchmark_T10_P500000_CountTask-4       432            376            -12.96%
Benchmark_T1000_P500-4                  32184          27190          -15.52%
Benchmark_T1000_P500000_CountTask-4     32244          27255          -15.47%

benchmark                               old bytes     new bytes     delta
Benchmark_T1000_P5000_Matches-4         1818368       1754257       -3.53%
Benchmark_T1000_P5000_NoMatches-4       1818384       1753251       -3.58%
Benchmark_T100_P5000_Matches-4          1414272       1407353       -0.49%
Benchmark_T100_P5000_NoMatches-4        1414289       1407314       -0.49%
Benchmark_T10_P5000_Matches-4           1373113       1372263       -0.06%
Benchmark_T10_P5000_NoMatches-4         1373109       1372239       -0.06%
Benchmark_T10_P500_CountTask-4          157380        155454        -1.22%
Benchmark_T10_P500000_CountTask-4       149611025     149609208     -0.00%
Benchmark_T1000_P500-4                  1433664       1274000       -11.14%
Benchmark_T1000_P500000_CountTask-4     150887360     150728200     -0.11%
```